% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_list.R
\name{lists}
\alias{lists}
\alias{as_list}
\alias{as_list2}
\title{Vaiations of \code{as.list} and \code{do.call}.}
\usage{
as_list(x, all.names = TRUE)

as_list2(x, ...)
}
\description{
Vaiations of \code{as.list} and \code{do.call}.
}
