% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_fun_file.R
\name{new_fun_file}
\alias{new_fun_file}
\title{Variant of \code{usethis::use_r}.}
\usage{
new_fun_file(fn, path = NULL, overwrite = TRUE, open = rlang::is_interactive())
}
\description{
Variant of \code{usethis::use_r}.
}
