% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split2.R
\name{str_split2}
\alias{str_split2}
\alias{str_split_i2}
\title{String Splitting Variations}
\usage{
str_split2(string, pattern, n, simplify = TRUE)

str_split_i2(string, pattern, i, simplify = TRUE)
}
\description{
String Splitting Variations
}
