% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_empty.R
\name{test_empty}
\alias{test_empty}
\alias{is_empty}
\alias{is_scalar_empty}
\alias{are_empty}
\alias{not_empty}
\alias{all_empty}
\alias{not_all_empty}
\alias{\%|0|\%}
\alias{if_empty}
\title{Test for NULLs}
\usage{
is_empty(x)

is_scalar_empty(x)

are_empty(x)

not_empty(x)

all_empty(x)

not_all_empty(x)

x \%|0|\% y

if_empty(x, y = x)
}
\description{
Test for NULLs
}
