% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_missing.R
\name{test_missing}
\alias{test_missing}
\alias{is_missing}
\alias{are_missing}
\alias{not_missing}
\alias{all_missing}
\alias{not_all_missing}
\alias{if_missing}
\title{Test for NULLs}
\usage{
is_missing(x)

are_missing(x)

not_missing(x)

all_missing(x)

not_all_missing(x)

if_missing(x, y = x)
}
\description{
Test for NULLs
}
