% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_missing2.R
\name{test_missing2}
\alias{test_missing2}
\alias{is_missing2}
\alias{are_missing2}
\alias{not_missing2}
\alias{all_missing2}
\alias{not_all_missing2}
\alias{if_missing2}
\title{Test for NULLs}
\usage{
is_missing2(x)

are_missing2(x)

not_missing2(x)

all_missing2(x)

not_all_missing2(x)

if_missing2(x, y = x)
}
\description{
Test for NULLs
}
