% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_na.R
\name{test_na}
\alias{test_na}
\alias{is_na}
\alias{are_na}
\alias{not_na}
\alias{all_na}
\alias{not_all_na}
\alias{\%na\%}
\alias{if_na}
\title{Test for NULLs}
\usage{
is_na(x)

is_na(x)

are_na(x)

not_na(x)

all_na(x)

not_all_na(x)

x \%na\% y

if_na(x, y = x)
}
\description{
Test for NULLs
}
