% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_null.R
\name{test_null}
\alias{test_null}
\alias{is_null}
\alias{is_scalar_null}
\alias{are_null}
\alias{not_null}
\alias{all_null}
\alias{not_all_null}
\alias{\%||\%}
\alias{if_null}
\title{Test for NULLs}
\usage{
is_null()

is_scalar_null(x)

are_null(x)

not_null(x)

all_null(x)

not_all_null(x)

x \%||\% y

if_null(x, y = x)
}
\description{
Test for NULLs
}
