% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial2_compose2.R
\name{partial_compose}
\alias{partial_compose}
\alias{partial2}
\alias{compose2}
\alias{pf}
\alias{cf}
\alias{make_fn}
\title{Variants of \code{partial} and \code{compose}}
\usage{
partial2(x, as_expr = TRUE, fn_out = TRUE)

compose2(...)

pf(x, as_expr = TRUE, fn_out = TRUE)

cf(...)

make_fn(x)
}
\arguments{
\item{as_expr}{indicates the need for enexpr0}

\item{fn_out}{returns the full orignal function body.}
}
\description{
Variants of \code{partial} and \code{compose}
}
