% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_slice.R
\name{slicing}
\alias{slicing}
\alias{vec_slice_pos}
\alias{vec_slice_head}
\alias{vec_slice_tail}
\title{Title}
\usage{
vec_slice_pos(x, i, ..., rev = FALSE)

vec_slice_head(x, n)

vec_slice_tail(x, n)
}
\description{
Title
}
